# Projects/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Top Dog Arena NFT API
  description: |
    RESTful API for creating and managing NFTs on the XRPL (XRP Ledger) for Top Dog Arena baseball card marketplace.
    
    ## Features
    - Create single and batch NFTs on XRPL
    - Query NFT details and ownership
    - Health monitoring endpoints
    - Rate limiting and validation
    
    ## Rate Limits
    - General API: 100 requests per 15 minutes per IP
    - NFT Creation: 10 requests per hour per IP  
    - Batch Operations: 3 requests per hour per IP
    - Read Operations: 200 requests per 15 minutes per IP
    
    ## XRPL Integration
    This API integrates with the XRP Ledger to mint NFTs representing baseball cards and other collectibles.
    
  version: 1.0.0
  contact:
    name: Top Dog Arena
    url: https://topdogarena.com
    email: support@topdogarena.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server with full database integration
  - url: http://localhost:3000
    description: Development server base
  - url: https://api.topdogarena.com/api
    description: Production server
...
paths:
...
  /nft/create:
    post:
      tags:
        - NFT
      summary: Create a single NFT
      description: |
        Creates a single NFT on the XRPL. Requires XRPL wallet to be configured.
        
        **Rate Limit:** 10 requests per hour per IP
      operationId: createNFT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNFTRequest'
            example:
              name: "Mike Trout - Angels Superstar"
              description: "A legendary baseball card featuring Mike Trout in his Angels uniform."
              image: "https://topdogarena.com/images/cards/mike-trout-001.png"
              attributes:
                - trait_type: "Player"
                  value: "Mike Trout"
                - trait_type: "Team"
                  value: "Los Angeles Angels"
                - trait_type: "Position"
                  value: "Center Field"
                - trait_type: "Rarity"
                  value: "Legendary"
              external_url: "https://topdogarena.com/players/mike-trout"
              transferFee: 500
      responses:
        '201':
          description: NFT created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NFTCreatedResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitErrorResponse'
        '503':
          description: XRPL service not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /nft/batch-create:
    post:
      tags:
        - NFT
      summary: Create multiple NFTs in batch
      description: |
        Creates up to 10 NFTs in a single batch operation.
        
        **Rate Limit:** 3 requests per hour per IP
      operationId: batchCreateNFTs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchCreateNFTRequest'
      responses:
        '201':
          description: All NFTs created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchNFTResponse'
        '207':
          description: Partial success - some NFTs failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchNFTResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitErrorResponse'
  /nft/{nftId}:
    get:
      tags:
        - NFT
      summary: Get NFT details by ID
      description: Retrieve detailed information about a specific NFT
      operationId: getNFTById
      parameters:
        - name: nftId
          in: path
          required: true
          description: 64-character hexadecimal NFT ID
          schema:
            type: string
            pattern: '^[0-9A-Fa-f]{64}$'
            example: "000B013A95F14B0044F78A264E41713C64B5F89242540EE208C3098E00000D65"
      responses:
        '200':
          description: NFT details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NFTDetailsResponse'
        '400':
          description: Invalid NFT ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: NFT not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /nft/account/{account}:
    get:
      tags:
        - NFT
      summary: Get NFTs owned by account
      description: Retrieve all NFTs owned by a specific XRPL account
      operationId: getAccountNFTs
      parameters:
        - name: account
          in: path
          required: true
          description: XRPL account address
          schema:
            type: string
            pattern: '^r[1-9A-HJ-NP-Za-km-z]{25,34}$'
            example: "rN7n7otQDd6FczFgLdSqtcsAUxDkw6fzRH"
        - name: limit
          in: query
          description: Number of results to return (1-100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: marker
          in: query
          description: Pagination marker for next page
          schema:
            type: string
      responses:
        '200':
          description: Account NFTs retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountNFTsResponse'
        '400':
          description: Invalid account address
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    CreateNFTRequest:
      type: object
      required:
        - name
        - description
        - image
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: "Mike Trout - Angels Superstar"
        description:
          type: string
          minLength: 1
          maxLength: 1000
          example: "A legendary baseball card featuring Mike Trout"
        image:
          type: string
          format: uri
          example: "https://topdogarena.com/images/cards/mike-trout-001.png"
        attributes:
          type: array
          items:
            type: object
            properties:
              trait_type:
                type: string
              value:
                oneOf:
                  - type: string
                  - type: number
          example:
            - trait_type: "Player"
              value: "Mike Trout"
            - trait_type: "Team"
              value: "Los Angeles Angels"
        external_url:
          type: string
          format: uri
        animation_url:
          type: string
          format: uri
        transferFee:
          type: integer
          minimum: 0
          maximum: 50000
          description: Transfer fee in 1/100,000th units (10000 = 10%)
        recipient:
          type: string
          pattern: '^r[1-9A-HJ-NP-Za-km-z]{25,34}$'
          description: XRPL address to mint NFT to
        flags:
          type: integer
          description: NFT flags (8 = transferable)
    BatchCreateNFTRequest:
      type: object
      required:
        - nfts
      properties:
        nfts:
          type: array
          minItems: 1
          maxItems: 10
          items:
            $ref: '#/components/schemas/CreateNFTRequest'
    NFTCreatedResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            nftId:
              type: string
              example: "000B013A95F14B0044F78A264E41713C64B5F89242540EE208C3098E00000D65"
            txHash:
              type: string
              example: "E3FE6EA3D48F0C2B639448020EA4F03D4F4F8BDBD2397B43FE71618E38D4B277"
            fee:
              type: string
              example: "12"
            metadata:
              $ref: '#/components/schemas/NFTMetadata'
        message:
          type: string
          example: "NFT created successfully"
    BatchNFTResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            successful:
              type: array
              items:
                type: object
                properties:
                  index:
                    type: integer
                  nftId:
                    type: string
                  txHash:
                    type: string
                  fee:
                    type: string
                  metadata:
                    $ref: '#/components/schemas/NFTMetadata'
            failed:
              type: array
              items:
                type: object
                properties:
                  index:
                    type: integer
                  error:
                    type: string
                  nftData:
                    type: object
            summary:
              type: object
              properties:
                total:
                  type: integer
                successful:
                  type: integer
                failed:
                  type: integer
        message:
          type: string
    NFTDetailsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            nft_id:
              type: string
            ledger_index:
              type: integer
            owner:
              type: string
            is_burned:
              type: boolean
            flags:
              type: integer
            transfer_fee:
              type: integer
            issuer:
              type: string
            nft_taxon:
              type: integer
            nft_serial:
              type: integer
            uri:
              type: string
              description: Hex-encoded metadata URI
        message:
          type: string
    AccountNFTsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            account:
              type: string
            nfts:
              type: array
              items:
                type: object
                properties:
                  Flags:
                    type: integer
                  Issuer:
                    type: string
                  NFTokenID:
                    type: string
                  NFTokenTaxon:
                    type: integer
                  TransferFee:
                    type: integer
                  URI:
                    type: string
                  nft_serial:
                    type: integer
            count:
              type: integer
        message:
          type: string
...
    NFTMetadata:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        image:
          type: string
        attributes:
          type: array
          items:
            type: object
            properties:
              trait_type:
                type: string
              value:
                oneOf:
                  - type: string
                  - type: number
        external_url:
          type: string
        animation_url:
          type: string
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
    ValidationErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Validation failed"
        message:
          type: string
          example: "Invalid request data"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
        timestamp:
          type: string
          format: date-time
    RateLimitErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Too many requests"
        message:
          type: string
          example: "Rate limit exceeded. Please try again later."
        retryAfter:
          type: string
          example: "15 minutes"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated requests (coming soon)
security: []